+++ backend/Models/Dto/GraphEdgeUpdateDto.cs
@@ -2,5 +2,5 @@ namespace Backend.Models.Dto;
 
 public class GraphEdgeUpdateDto
 {
-    public int? ImportanceToParent { get; set; }
+    public decimal? ImportanceToParent { get; set; }
 }
